requirements.txt
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
Flask==1.1.2
flask-heroku==0.1.9
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-dotenv==0.19.2
requests==2.25.1
SQLAlchemy==1.4.27
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
